🔥 chore: move src/api/github directory to src/services/github
